Removed console output and type checking
git-svn-id: https://xbox/svnroot/Qorus/trunk@4838 e9028eec-4831-4325-a917-4dc279fcb47c

+++ src/js/collections/notifications.js
@@ -63,7 +63,8 @@ define(function (require) {
         this.trigger(sprintf('cleared:%s', group));
       } else {
         _(this.models).each(function (model) {
-          model.destroy({ silent: true });
+          if (model instanceof Backbone.Model)
+            model.destroy({ silent: true });
         });
         this.reset();
       }
